# backend/src/schemas/usuario.py
import uuid
from datetime import datetime

from pydantic import BaseModel, ConfigDict, EmailStr, Field


class RegistroRequest(BaseModel):
    email: EmailStr
    username: str = Field(min_length=3, max_length=30, pattern=r"^[a-zA-Z0-9_-]+$")
    display_name: str = Field(min_length=1, max_length=80)
    password: str = Field(min_length=8)


class LoginRequest(BaseModel):
    email: EmailStr
    password: str


class Usuario(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: uuid.UUID
    username: str
    display_name: str
    bio: str | None
    avatar_url: str | None
    created_at: datetime


class PerfilUsuario(Usuario):
    total_opinioes: int


class AuthResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    usuario: Usuario


class UsuarioBusca(Usuario):
    ja_seguido: bool


class UsuarioBuscaListaResponse(BaseModel):
    items: list[UsuarioBusca]
    page: int
    page_size: int
    total: int


class UsuarioListaResponse(BaseModel):
    items: list[Usuario]
    page: int
    page_size: int
    total: int